requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
pypdf>=3.17.0
python-dotenv>=1.0.0
streamlit>=1.28.0
faiss-cpu>=1.7.0 